refs #13923 Move common methods to shop_notice & added plan_year template for SHOP
