+++ app/notices/shop_notices/conversion_renewal_notice.rb
@@ -1,4 +1,4 @@
-class ShopNotices::ConversionRenewalNotice < ShopNotices::RenewalGroupNotice
+class ShopNotices::ConversionRenewalNotice < ShopNotice
 
   def deliver
     build
